+++ Erebus/source.cpp
@@ -11,13 +11,16 @@
 #include "Controls.h"
 #include "ParticleImport.h"
 
-#include "GamePlay.h"
 #include "Menu.h"
 #include "CollisionChecker.h"
 #include "RayCollider.h"
 #include "SoundEngine.h"
 #include "WorkQueue.h"
+#include "CollisionHandler.h"
 #include "Frustum.h"
+#include "AGI.h"
+#include "NetworkController.hpp"
+#include "LuaBinds.h"
 
 #define MAX_TRANSFORMS 100
 #define MAX_ANIMATIONS 100
@@ -141,11 +144,11 @@ DWORD WINAPI update( LPVOID args )
 		{
 			double deltaTime = counter.getDeltaTime();
 
-			luaBinds.update( data->controls, deltaTime );
+			luaBinds.update( data->controls, (float)deltaTime );
 			data->workQueue->execute();
 
 			for( int i=0; i<data->particleSystems->size(); i++ )
-				data->particleSystems->at(i)->update( deltaTime );
+				data->particleSystems->at(i)->update( (float)deltaTime );
 
 			collisionHandler.checkCollisions();
 
@@ -153,10 +156,11 @@ DWORD WINAPI update( LPVOID args )
 				+ "\nVRAM: " + std::to_string(counter.getVramUsage()) + " MB" 
 				+ "\nRAM: " + std::to_string(counter.getRamUsage()) + " MB";
 			data->engine->print(fps, 0.0f, 0.0f);
+			//data->engine->print(data->soundEngine->getDbgTxt(), 350, 0, 0.7);
 
 			for( int i=0; i<boundAnimations; i++ )
 			{
-				animationData[i].dt = deltaTime;
+				animationData[i].dt = (float)deltaTime;
 				//data->allAnimations[i].update(deltaTime);
 				data->workQueue->add( updateAnimation, &animationData[i] );
 			}
@@ -323,7 +327,7 @@ int main()
 			engine.draw(&camera);
 
 #ifdef _DEBUG
-			assets.checkHotload(deltaTime);
+			assets.checkHotload((float)deltaTime);
 #endif // DEBUG
 		}
 	}
